--- core/src/test/java/dev/keva/core/server/AbstractServerTest.java
@@ -1,1209 +1,0 @@
-package dev.keva.core.server;
-
-import dev.keva.core.command.impl.string.GetEx;
-import lombok.val;
-import org.junit.jupiter.api.Test;
-import org.junit.jupiter.api.Timeout;
-import redis.clients.jedis.Jedis;
-import redis.clients.jedis.JedisPubSub;
-import redis.clients.jedis.exceptions.JedisDataException;
-
-import java.util.*;
-import java.util.concurrent.CompletableFuture;
-import java.util.concurrent.ExecutionException;
-
-import redis.clients.jedis.params.GetExParams;
-import redis.clients.jedis.params.StrAlgoLCSParams;
-import redis.clients.jedis.params.ZAddParams;
-import redis.clients.jedis.resps.LCSMatchResult;
-
-import static org.junit.jupiter.api.Assertions.*;
-
-public abstract class AbstractServerTest {
-    static Jedis jedis;
-    static Server server;
-    static Jedis subscriber;
-
-    @Test
-    void ping() {
-        try {
-            val pong = jedis.ping();
-            assertTrue("PONG".contentEquals(pong));
-            assertEquals("PONG", pong);
-        } catch (Exception e) {
-            fail(e);
-        }
-    }
-
-    @Test
-    void info() {
-        try {
-            val info = jedis.info();
-            assertNotNull(info);
-        } catch (Exception e) {
-            fail(e);
-        }
-    }
-
-    @Test
-    void flush() {
-        try {
-            val setAbc = jedis.set("abc", "123");
-            assertEquals("OK", setAbc);
-            val flush = jedis.flushDB();
-            assertEquals("OK", flush);
-            val getAbc = jedis.get("abc");
-            assertNull(getAbc);
-        } catch (Exception e) {
-            fail(e);
-        }
-    }
-
-    @Test
-    void getSetNull() {
-        try {
-            val getNull = jedis.get("anotherkey");
-            assertNull(getNull);
-        } catch (Exception e) {
-            fail(e);
-        }
-    }
-
-    @Test
-    void getSet() {
-        try {
-            val setAbc = jedis.set("abc", "123");
-            val getAbc = jedis.get("abc");
-            assertEquals("OK", setAbc);
-            assertEquals("123", getAbc);
-        } catch (Exception e) {
-            fail(e);
-        }
-    }
-
-    @Test
-    void getSetEmptyString() {
-        try {
-            val setEmpty = jedis.set("empty", "");
-            val getEmpty = jedis.get("empty");
-            assertEquals("OK", setEmpty);
-            assertEquals("", getEmpty);
-        } catch (Exception e) {
-            fail(e);
-        }
-    }
-
-    @Test
-    void del() {
-        try {
-            String setAbc = jedis.set("abc", "123");
-            val getAbc = jedis.get("abc");
-            val delAbc = jedis.del("abc");
-            val getAbcNull = jedis.get("abc");
-            assertEquals("OK", setAbc);
-            assertEquals("123", getAbc);
-            assertEquals(1, delAbc);
-            assertNull(getAbcNull);
-        } catch (Exception e) {
-            fail(e);
-        }
-    }
-
-    @Test
-    void wrongTypeSet() {
-        try {
-            val setAbc = jedis.set("abc", "123");
-            assertEquals("OK", setAbc);
-            jedis.lpush("abc", "123");
-        } catch (Exception e) {
-            assertEquals(JedisDataException.class, e.getClass());
-        }
-    }
-
-    @Test
-    void wrongTypeSet2() {
-        try {
-            val setAbc = jedis.lpush("abc", "123");
-            assertEquals(1, setAbc);
-            jedis.hset("abc", "key", "val");
-        } catch (Exception e) {
-            assertEquals(JedisDataException.class, e.getClass());
-        }
-    }
-
-    @Test
-    void wrongTypeSet3() {
-        try {
-            val setAbc = jedis.hset("abc", "key", "val");
-            assertEquals(1, setAbc);
-            jedis.lpush("abc", "123");
-        } catch (Exception e) {
-            assertEquals(JedisDataException.class, e.getClass());
-        }
-    }
-
-    @Test
-    void transaction() {
-        try {
-            val transaction = jedis.multi();
-            transaction.set("abc", "123");
-            transaction.set("def", "456");
-            val exec = transaction.exec();
-            assertNotNull(exec);
-            assertEquals(2, exec.size());
-            assertEquals("OK", exec.get(0));
-            assertEquals("OK", exec.get(1));
-
-            val getAbc = jedis.get("abc");
-            val getDef = jedis.get("def");
-            assertEquals("123", getAbc);
-            assertEquals("456", getDef);
-        } catch (Exception e) {
-            fail(e);
-        }
-    }
-
-    @Test
-    void transactionWatch() {
-        try {
-            jedis.watch("abc");
-            val transaction = jedis.multi();
-            transaction.set("abc", "123");
-            transaction.set("abc", "456");
-            val exec = transaction.exec();
-            assertNotNull(exec);
-            assertEquals(2, exec.size());
-            assertEquals("OK", exec.get(0));
-            assertEquals("OK", exec.get(1));
-
-            val getAbc = jedis.get("abc");
-            assertEquals("456", getAbc);
-        } catch (Exception e) {
-            fail(e);
-        }
-    }
-
-    @Test
-    void getSetExpire() {
-        try {
-            val setAbc = jedis.set("abc", "123");
-            String getAbc = jedis.get("abc");
-            val expireAbc = jedis.expire("abc", 1L);
-
-            assertEquals("OK", setAbc);
-            assertEquals("123", getAbc);
-            assertEquals(1, expireAbc);
-            Thread.sleep(500);
-            getAbc = jedis.get("abc");
-            assertEquals("123", getAbc);
-            Thread.sleep(501);
-            val getAbcNull = jedis.get("abc");
-            assertNull(getAbcNull);
-        } catch (Exception e) {
-            fail(e);
-        }
-    }
-
-    @Test
-    void updateExpire() {
-        try {
-            val setAbc = jedis.set("abc", "123");
-            String getAbc = jedis.get("abc");
-            Long expireAbc = jedis.expire("abc", 1L);
-
-            assertEquals("OK", setAbc);
-            assertEquals("123", getAbc);
-            assertEquals(1, expireAbc);
-
-            Thread.sleep(500);
-            getAbc = jedis.get("abc");
-            expireAbc = jedis.expire("abc", 1L);
-            assertEquals(1, expireAbc);
-            assertEquals("123", getAbc);
-
-            Thread.sleep(501);
-            getAbc = jedis.get("abc");
-            assertEquals("123", getAbc);
-
-            Thread.sleep(1001);
-            val getAbcNull = jedis.get("abc");
-            assertNull(getAbcNull);
-        } catch (Exception e) {
-            fail(e);
-        }
-    }
-
-    @Test
-    void getSetExpireAt() {
-        try {
-            val setAbc = jedis.set("abc", "123");
-            String getAbc = jedis.get("abc");
-            val oneSecondLaterTime = System.currentTimeMillis() + 1000;
-            val expireAbc = jedis.expireAt("abc", oneSecondLaterTime);
-
-            assertEquals("OK", setAbc);
-            assertEquals("123", getAbc);
-            assertEquals(1, expireAbc);
-            Thread.sleep(500);
-            getAbc = jedis.get("abc");
-            assertEquals("123", getAbc);
-            Thread.sleep(501);
-            val getAbcNull = jedis.get("abc");
-            assertNull(getAbcNull);
-        } catch (Exception e) {
-            fail(e);
-        }
-    }
-
-    @Test
-    void setAfterExpireAt() {
-        try {
-            String setAbc = jedis.set("abc", "123");
-            String getAbc = jedis.get("abc");
-            val oneSecondLaterTime = System.currentTimeMillis() + 1000;
-            val expireAbc = jedis.expireAt("abc", oneSecondLaterTime);
-
-            assertEquals("OK", setAbc);
-            assertEquals("123", getAbc);
-            assertEquals(1, expireAbc);
-            Thread.sleep(500);
-            getAbc = jedis.get("abc");
-            assertEquals("123", getAbc);
-            setAbc = jedis.set("abc", "456");
-            assertEquals("OK", setAbc);
-            Thread.sleep(501);
-            getAbc = jedis.get("abc");
-            assertEquals("456", getAbc);
-        } catch (Exception e) {
-            fail(e);
-        }
-    }
-
-    @Test
-    @Timeout(5)
-    void getSetLongString() {
-        try {
-            final int aKB = 1024;
-            String testStr = String.join("", Collections.nCopies(aKB, "a"));
-            String setAbc = jedis.set("abc", testStr);
-            String getAbc = jedis.get("abc");
-            assertEquals("OK", setAbc);
-            assertEquals(testStr, getAbc);
-        } catch (Exception e) {
-            fail(e);
-        }
-    }
-
-    @Test
-    @Timeout(5)
-    void getSetLongKeyString() {
-        try {
-            final int aKB = 1026;
-            final String testStr = String.join("", Collections.nCopies(aKB, "a"));
-            final String setAbc = jedis.set(testStr, "123");
-            final String getAbc = jedis.get(testStr);
-            assertEquals("OK", setAbc);
-            assertEquals("123", getAbc);
-        } catch (Exception e) {
-            fail(e);
-        }
-    }
-
-    @Test
-    @Timeout(30)
-    void pubsub() throws ExecutionException, InterruptedException {
-        CompletableFuture<String> future = new CompletableFuture<>();
-        new Thread(() -> subscriber.subscribe(new JedisPubSub() {
-            @Override
-            public void onMessage(String channel, String message) {
-                future.complete(message);
-            }
-
-            @Override
-            public void onSubscribe(String channel, int subscribedChannels) {
-                jedis.publish("test", "Test message");
-            }
-        }, "test")).start();
-        val message = future.get();
-        assertEquals("Test message", message);
-    }
-
-    @Test
-    @Timeout(5)
-    void rename() {
-        try {
-            final String initKey = "Key";
-            final String initVal = "Val";
-            final String newKey = "Nkey";
-            final String renameBeforeSet = jedis.rename(initKey, newKey);
-            jedis.set(initKey, initVal);
-            final String renameAfterSet = jedis.rename(initKey, newKey);
-            final String getAfterRename = jedis.get(newKey);
-            assertEquals("ERR unknown key", renameBeforeSet);
-            assertEquals("OK", renameAfterSet);
-            assertEquals(initVal, getAfterRename);
-        } catch (Exception e) {
-            fail(e);
-        }
-    }
-
-    @Test
-    @Timeout(5)
-    void renameWithExpire() {
-        try {
-            val initKey = "Key";
-            val initVal = "Val";
-            val newKey = "Nkey";
-            val renameBeforeSet = jedis.rename(initKey, newKey);
-            jedis.set(initKey, initVal);
-            jedis.expireAt(initKey, System.currentTimeMillis() + 500);
-            val renameAfterSet = jedis.rename(initKey, newKey);
-            String getAfterRename = jedis.get(newKey);
-            assertEquals("ERR unknown key", renameBeforeSet);
-            assertEquals("OK", renameAfterSet);
-            assertEquals(initVal, getAfterRename);
-            Thread.sleep(501);
-            getAfterRename = jedis.get(newKey);
-            assertNull(getAfterRename);
-        } catch (Exception e) {
-            fail(e);
-        }
-    }
-
-    @Test
-    void getdel() {
-        try {
-            val setAbc = jedis.set("abc", "123");
-            val getdelAbc = jedis.getDel("abc");
-            val getAbcAgain = jedis.get("abc");
-            assertEquals("OK", setAbc);
-            assertEquals("123", getdelAbc);
-            assertNull(getAbcAgain);
-        } catch (Exception e) {
-            fail(e);
-        }
-    }
-
-    @Test
-    void append() {
-        try {
-            val append1 = jedis.append("1", "Hello");
-            val append2 = jedis.append("1", " World");
-            val get = jedis.get("1");
-            assertEquals(5, append1);
-            assertEquals(11, append2);
-            assertEquals("Hello World", get);
-        } catch (Exception e) {
-            fail(e);
-        }
-    }
-
-    @Test
-    void incr() {
-        // with exist key
-        String res = jedis.set("1to2", "1");
-        assertEquals("OK", res);
-        Long newVal = jedis.incr("1to2");
-        assertEquals(2, newVal);
-        String val = jedis.get("1to2");
-        assertEquals("2", val);
-
-        // with non exist key
-        res = jedis.get("0to1");
-        assertNull(res);
-        newVal = jedis.incr("0to1");
-        assertEquals(1, newVal);
-        val = jedis.get("0to1");
-        assertEquals("1", val);
-
-        // with wrong key type
-        res = jedis.set("wrong", "type");
-        assertEquals("OK", res);
-        assertThrows(JedisDataException.class, () -> jedis.incr("wrong"));
-    }
-
-    @Test
-    void incrBy() {
-        // with exist key
-        String res = jedis.set("1to5", "1");
-        assertEquals("OK", res);
-        Long newVal = jedis.incrBy("1to5", 4);
-        assertEquals(5, newVal);
-        String val = jedis.get("1to5");
-        assertEquals("5", val);
-
-        // with non exist key
-        res = jedis.get("0to10");
-        assertNull(res);
-        newVal = jedis.incrBy("0to10", 10);
-        assertEquals(10, newVal);
-        val = jedis.get("0to10");
-        assertEquals("10", val);
-
-        // with wrong key type
-        res = jedis.set("wrong", "type");
-        assertEquals("OK", res);
-        assertThrows(JedisDataException.class, () -> jedis.incrBy("wrong", 10));
-    }
-
-    @Test
-    void hsetGet() {
-        try {
-            val hset = jedis.hset("test", "key", "val");
-            val hget = jedis.hget("test", "key");
-            assertEquals(1, hset);
-            assertEquals("val", hget);
-        } catch (Exception e) {
-            fail(e);
-        }
-    }
-
-    @Test
-    void hdel() {
-        try {
-            val hset = jedis.hset("test", "key", "val");
-            val hdel = jedis.hdel("test", "key");
-            assertEquals(1, hset);
-            assertEquals(1, hdel);
-        } catch (Exception e) {
-            fail(e);
-        }
-    }
-
-    @Test
-    void hexists() {
-        try {
-            val hset = jedis.hset("test", "key", "val");
-            val hexists = jedis.hexists("test", "key");
-            assertEquals(1, hset);
-            assertEquals(true, hexists);
-        } catch (Exception e) {
-            fail(e);
-        }
-    }
-
-    @Test
-    void hvals() {
-        try {
-            val hset = jedis.hset("test", "key", "val");
-            val hvals = jedis.hvals("test");
-            assertEquals(1, hset);
-            assertEquals(1, hvals.size());
-            assertEquals("val", hvals.get(0));
-        } catch (Exception e) {
-            fail(e);
-        }
-    }
-
-    @Test
-    void hkeys() {
-        try {
-            val hset = jedis.hset("test", "key", "val");
-            val hkeys = jedis.hkeys("test");
-            assertEquals(1, hset);
-            assertEquals(1, hkeys.size());
-            assertEquals("key", hkeys.toArray()[0]);
-        } catch (Exception e) {
-            fail(e);
-        }
-    }
-
-    @Test
-    void hlen() {
-        try {
-            val hset = jedis.hset("test", "key", "val");
-            val hlen = jedis.hlen("test");
-            assertEquals(1, hset);
-            assertEquals(1, hlen);
-        } catch (Exception e) {
-            fail(e);
-        }
-    }
-
-    @Test
-    void hstrlen() {
-        try {
-            val hset = jedis.hset("test", "key", "val");
-            val hstrlen = jedis.hstrlen("test", "key");
-            assertEquals(1, hset);
-            assertEquals(3, hstrlen);
-        } catch (Exception e) {
-            fail(e);
-        }
-    }
-
-    @Test
-    void hgetAll() {
-        try {
-            val hset = jedis.hset("test", "key", "val");
-            val hgetAll = jedis.hgetAll("test");
-            assertEquals(1, hset);
-            assertEquals(1, hgetAll.size());
-            assertEquals("val", hgetAll.get("key"));
-        } catch (Exception e) {
-            fail(e);
-        }
-    }
-
-    @Test
-    void lpush() {
-        try {
-            val lpush = jedis.lpush("test", "val");
-            assertEquals(1, lpush);
-        } catch (Exception e) {
-            fail(e);
-        }
-    }
-
-    @Test
-    void rpush() {
-        try {
-            val rpush = jedis.rpush("test", "val");
-            assertEquals(1, rpush);
-        } catch (Exception e) {
-            fail(e);
-        }
-    }
-
-    @Test
-    void lpop() {
-        try {
-            val lpush = jedis.lpush("test", "val");
-            val lpop = jedis.lpop("test");
-            assertEquals(1, lpush);
-            assertEquals("val", lpop);
-        } catch (Exception e) {
-            fail(e);
-        }
-    }
-
-    @Test
-    void rpop() {
-        try {
-            val lpush = jedis.lpush("test", "val");
-            val rpop = jedis.rpop("test");
-            assertEquals(1, lpush);
-            assertEquals("val", rpop);
-        } catch (Exception e) {
-            fail(e);
-        }
-    }
-
-    @Test
-    void llen() {
-        try {
-            val lpush = jedis.lpush("test", "val");
-            val llen = jedis.llen("test");
-            assertEquals(1, lpush);
-            assertEquals(1, llen);
-        } catch (Exception e) {
-            fail(e);
-        }
-    }
-
-    @Test
-    void lrange() {
-        try {
-            val lpush = jedis.lpush("test", "val");
-            val lrange = jedis.lrange("test", 0, 1);
-            assertEquals(1, lpush);
-            assertEquals(1, lrange.size());
-            assertEquals("val", lrange.get(0));
-        } catch (Exception e) {
-            fail(e);
-        }
-    }
-
-    @Test
-    void lindex() {
-        try {
-            val lpush = jedis.lpush("test", "val");
-            val lindex = jedis.lindex("test", 0);
-            assertEquals(1, lpush);
-            assertEquals("val", lindex);
-        } catch (Exception e) {
-            fail(e);
-        }
-    }
-
-    @Test
-    void lset() {
-        try {
-            val lpush = jedis.lpush("test", "val");
-            val lset = jedis.lset("test", 0, "newVal");
-            assertEquals(1, lpush);
-            assertEquals("OK", lset);
-        } catch (Exception e) {
-            fail(e);
-        }
-    }
-
-    @Test
-    void lrem() {
-        try {
-            val lpush = jedis.lpush("test", "val");
-            val lrem = jedis.lrem("test", 0, "val");
-            assertEquals(1, lpush);
-            assertEquals(1, lrem);
-        } catch (Exception e) {
-            fail(e);
-        }
-    }
-
-    @Test
-    void sadd() {
-        try {
-            val sadd = jedis.sadd("test", "val");
-            assertEquals(1, sadd);
-        } catch (Exception e) {
-            fail(e);
-        }
-    }
-
-    @Test
-    void smembers() {
-        try {
-            val sadd = jedis.sadd("test", "val");
-            val smembers = jedis.smembers("test");
-            assertEquals(1, sadd);
-            assertEquals(1, smembers.size());
-            assertEquals("val", smembers.toArray()[0]);
-        } catch (Exception e) {
-            fail(e);
-        }
-    }
-
-    @Test
-    void sismember() {
-        try {
-            val sadd = jedis.sadd("test", "val");
-            val sismember = jedis.sismember("test", "val");
-            assertEquals(1, sadd);
-            assertEquals(true, sismember);
-        } catch (Exception e) {
-            fail(e);
-        }
-    }
-
-    @Test
-    void scard() {
-        try {
-            val sadd = jedis.sadd("test", "val");
-            val scard = jedis.scard("test");
-            assertEquals(1, sadd);
-            assertEquals(1, scard);
-        } catch (Exception e) {
-            fail(e);
-        }
-    }
-
-    @Test
-    void sdiff() {
-        try {
-            val sadd = jedis.sadd("test", "val");
-            val sdiff = jedis.sdiff("test", "test2");
-            assertEquals(1, sadd);
-            assertEquals(1, sdiff.size());
-        } catch (Exception e) {
-            fail(e);
-        }
-    }
-
-    @Test
-    void sinter() {
-        try {
-            val sadd = jedis.sadd("test", "val");
-            val sadd2 = jedis.sadd("test2", "val");
-            val sinter = jedis.sinter("test", "test2");
-            assertEquals(1, sadd);
-            assertEquals(1, sadd2);
-            assertEquals(1, sinter.size());
-        } catch (Exception e) {
-            fail(e);
-        }
-    }
-
-    @Test
-    void sunion() {
-        try {
-            val sadd = jedis.sadd("test", "val");
-            val sunion = jedis.sunion("test", "test2");
-            assertEquals(1, sadd);
-            assertEquals(1, sunion.size());
-            assertEquals("val", sunion.toArray()[0]);
-        } catch (Exception e) {
-            fail(e);
-        }
-    }
-
-    @Test
-    void smove() {
-        try {
-            val sadd = jedis.sadd("test", "val");
-            val smove = jedis.smove("test", "test2", "val");
-            assertEquals(1, sadd);
-            assertEquals(1, smove);
-        } catch (Exception e) {
-            fail(e);
-        }
-    }
-
-    @Test
-    void srem() {
-        try {
-            val sadd = jedis.sadd("test", "val");
-            val srem = jedis.srem("test", "val");
-            assertEquals(1, sadd);
-            assertEquals(1, srem);
-        } catch (Exception e) {
-            fail(e);
-        }
-    }
-
-    @Test
-    void strlen() {
-        try {
-            val set1 = jedis.set("mykey", "Hello World");
-            assertEquals("OK", set1);
-            Long strlen1 = jedis.strlen("mykey");
-            Long strlen2 = jedis.strlen("nonexisting");
-            assertEquals(11, strlen1);
-            assertEquals(0, strlen2);
-        } catch (Exception e) {
-            fail(e);
-        }
-    }
-
-    @Test
-    void getset() {
-        try {
-            val set1 = jedis.set("mykey", "Hello");
-            assertEquals("OK", set1);
-            val getset1 = jedis.getSet("mykey", "World");
-            assertEquals("Hello", getset1);
-            val get1 = jedis.get("mykey");
-            assertEquals("World", get1);
-        } catch (Exception e) {
-            fail(e);
-        }
-    }
-
-    @Test
-    void mget() {
-        try {
-            val set1 = jedis.set("mykey1", "Hello");
-            assertEquals("OK", set1);
-            val set2 = jedis.set("mykey2", "World");
-            assertEquals("OK", set2);
-            val mget1 = jedis.mget("mykey1", "mykey2", "nonexistingkey");
-            assertEquals(mget1.size(), 3);
-            assertEquals(mget1, Arrays.asList("Hello", "World", null));
-        } catch (Exception e) {
-            fail(e);
-        }
-    }
-
-    @Test
-    void setrange() {
-        try {
-            val set1 = jedis.set("key1", "Hello World");
-            assertEquals("OK", set1);
-            val setrange1 = jedis.setrange("key1", 6, "Keva");
-            assertEquals(11, setrange1);
-            val get1 = jedis.get("key1");
-            assertEquals("Hello Kevad", get1);
-
-            val setrange2 = jedis.setrange("key2", 5, "Kevas");
-            assertEquals(10, setrange2);
-            val get2 = jedis.get("key2");
-            assertEquals("\u0000\u0000\u0000\u0000\u0000Kevas", get2);
-
-            val setrange3 = jedis.setrange("key2", 0, "Ke");
-            assertEquals(10, setrange3);
-            val get3 = jedis.get("key2");
-            assertEquals("Ke\u0000\u0000\u0000Kevas", get3);
-
-        } catch (Exception e) {
-            fail(e);
-        }
-    }
-
-    @Test
-    void zaddWithXXAndNXErrs() {
-        assertThrows(JedisDataException.class, () ->
-                jedis.zadd("zset", 1.0, "val", new ZAddParams().xx().nx()));
-    }
-
-    @Test
-    void zaddSingleWithNxAndGtErrs() {
-        assertThrows(JedisDataException.class, () ->
-                jedis.zadd("zset", 1.0, "val", new ZAddParams().gt().nx()));
-    }
-
-    @Test
-    void zaddSingleWithNxAndLtErrs() {
-        assertThrows(JedisDataException.class, () ->
-                jedis.zadd("zset", 1.0, "val", new ZAddParams().lt().nx()));
-    }
-
-    @Test
-    void zaddSingleWithGtAndLtErrs() {
-        assertThrows(JedisDataException.class, () ->
-                jedis.zadd("zset", 1.0, "val", new ZAddParams().lt().gt()));
-    }
-
-    @Test
-    void zaddSingleWithoutOptions() {
-        try {
-            Long result = jedis.zadd("zset", 1.0, "val");
-            assertEquals(1, result);
-
-            result = jedis.zadd("zset", 1.0, "val");
-            assertEquals(0, result);
-        } catch (Exception e) {
-            fail(e);
-        }
-    }
-
-    @Test
-    void zaddMultipleWithoutOptions() {
-        try {
-            Map<String, Double> members = new HashMap<>();
-            int numMembers = 100;
-            for(int i=0; i<numMembers; ++i) {
-                members.put(Integer.toString(i), (double) i);
-            }
-            Long result = jedis.zadd("zset", members);
-            assertEquals(numMembers, result);
-
-            result = jedis.zadd("zset", members);
-            assertEquals(0, result);
-        } catch (Exception e) {
-            fail(e);
-        }
-    }
-
-    @Test
-    void zaddCh() {
-        try {
-            Long result = jedis.zadd("zset", 1.0, "mem", new ZAddParams().ch());
-            assertEquals(1, result);
-
-            result = jedis.zadd("zset", 1.0, "mem", new ZAddParams().ch());
-            assertEquals(0, result);
-
-            result = jedis.zadd("zset", 2.0, "mem", new ZAddParams().ch());
-            assertEquals(1, result);
-        } catch (Exception e) {
-            fail(e);
-        }
-    }
-
-    @Test
-    void zscoreNonExistingKey() {
-        val result = jedis.zscore("key", "mem");
-        assertNull(result);
-    }
-
-    @Test
-    void zscoreNonExistingMember() {
-        jedis.zadd("zset", 1.0, "mem");
-        val result = jedis.zscore("zset", "foo");
-        assertNull(result);
-    }
-
-    @Test
-    void zscoreExistingMember() {
-        jedis.zadd("zset", 1.0, "mem");
-        val result = jedis.zscore("zset", "mem");
-        assertEquals(result, 1.0);
-    }
-
-    @Test
-    void dumpAndRestore() {
-        try {
-            val set1 = jedis.set("key1", "Hello World");
-            assertEquals("OK", set1);
-            val dump1 = jedis.dump("key1");
-            assertNotNull(dump1);
-            val restore1 = jedis.restore("key2", 0L, dump1);
-            assertEquals("OK", restore1);
-            val key2 = jedis.get("key2");
-            assertEquals("Hello World", key2);
-        } catch (Exception e) {
-            fail(e);
-        }
-    }
-
-    @Test
-    void decr() {
-        try {
-            String set1 = jedis.set("mykey", "10");
-            assertEquals(set1, "OK");
-            Long decr1 = jedis.decr("mykey");
-            assertEquals(decr1, 9);
-            String set2 = jedis.set("errKey", "foobar");
-            assertThrows(JedisDataException.class, () -> jedis.decr("errKey"));
-
-        } catch (Exception e) {
-            fail(e);
-        }
-    }
-
-    @Test
-    void decrBy() {
-        try {
-            String set1 = jedis.set("mykey", "10");
-            assertEquals(set1, "OK");
-            Long decrby1 = jedis.decrBy("mykey", 5);
-            assertEquals(decrby1, 5);
-            Long decrby2 = jedis.decrBy("mykey", 10);
-            assertEquals(decrby2, -5);
-            String set2 = jedis.set("mykey2", "abc123");
-
-            assertThrows(JedisDataException.class, () -> jedis.decrBy("mykey2", 10));
-        } catch (Exception e) {
-            fail(e);
-        }
-    }
-
-    @Test
-    void type() {
-        try {
-            val set1 = jedis.set("key1", "Hello World");
-            assertEquals("OK", set1);
-            val type1 = jedis.type("key1");
-            assertEquals("string", type1);
-            val type2 = jedis.type("key2");
-            assertEquals("none", type2);
-            val hashSet = jedis.hset("key3", "field1", "value1");
-            assertEquals(1, hashSet.intValue());
-            val type3 = jedis.type("key3");
-            assertEquals("hash", type3);
-        } catch (Exception e) {
-            fail(e);
-        }
-    }
-
-    @Test
-    void getRange() {
-        try {
-            String set1 = jedis.set("mykey", "This is a string");
-            String getrange1 = jedis.getrange("mykey", 0, 3);
-            assertEquals("This", getrange1);
-            String getrange2 = jedis.getrange("mykey", -3, -1);
-            assertEquals("ing", getrange2);
-            String getrange3 = jedis.getrange("mykey", 0, -1);
-            assertEquals("This is a string", getrange3);
-            String getrange4 = jedis.getrange("mykey", 10, 100);
-            assertEquals("string", getrange4);
-            String getrange5 = jedis.getrange("mykey", 10, 5);
-            assertEquals("", getrange5);
-            String getrange6 = jedis.getrange("mykey", -10, 10);
-            assertEquals("s a s", getrange6);
-        } catch (Exception e) {
-            fail(e);
-        }
-    }
-
-    @Test
-    void incrByFloat() {
-        try {
-            String set1 = jedis.set("mykey", "10.50");
-            assertEquals(set1, "OK");
-            Double incrbyfloat1 = jedis.incrByFloat("mykey", 0.1);
-            assertEquals(incrbyfloat1, 10.6);
-            Double incrbyfloat2 = jedis.incrByFloat("mykey", -5);
-            assertEquals(incrbyfloat2, 5.6);
-            String set2 = jedis.set("mykey", "5.0e3");
-            assertEquals(set2, "OK");
-            Double incrbyfloat3 = jedis.incrByFloat("mykey", 2.0e2);
-            assertEquals(incrbyfloat3, 5200);
-            String set3 = jedis.set("mykey3", "abc");
-            assertEquals(set3, "OK");
-            assertThrows(JedisDataException.class, () -> jedis.incrByFloat("mykey3", 123));
-        } catch (Exception e) {
-            fail(e);
-        }
-    }
-
-    @Test
-    void mset() {
-        try {
-            String mset1 = jedis.mset("key1", "Hello", "key2", "World");
-            assertEquals("OK", mset1);
-            String get1 = jedis.get("key1");
-            assertEquals("Hello", get1);
-            String get2 = jedis.get("key2");
-            assertEquals("World", get2);
-        } catch (Exception e) {
-            fail(e);
-        }
-    }
-
-    @Test
-    void lcs1() {
-        try {
-            String mset1 = jedis.mset("key1", "ohmytext", "key2", "mynewtext");
-            assertEquals("OK", mset1);
-
-            LCSMatchResult lcs1 = jedis.strAlgoLCSKeys("key1", "key2", StrAlgoLCSParams.StrAlgoLCSParams().idx().withMatchLen());
-
-            // match string: mytext
-            assertEquals(lcs1.getLen(), 6);
-
-            assertEquals(lcs1.getMatches().get(0).getA().getStart(), 4);
-            assertEquals(lcs1.getMatches().get(0).getA().getEnd(), 7);
-            assertEquals(lcs1.getMatches().get(0).getB().getStart(), 5);
-            assertEquals(lcs1.getMatches().get(0).getB().getEnd(), 8);
-            assertEquals(lcs1.getMatches().get(0).getMatchLen(), 4);
-
-            assertEquals(lcs1.getMatches().get(1).getA().getStart(), 2);
-            assertEquals(lcs1.getMatches().get(1).getA().getEnd(), 3);
-            assertEquals(lcs1.getMatches().get(1).getB().getStart(), 0);
-            assertEquals(lcs1.getMatches().get(1).getB().getEnd(), 1);
-            assertEquals(lcs1.getMatches().get(1).getMatchLen(), 2);
-        } catch (Exception e) {
-            fail(e);
-        }
-    }
-
-    @Test
-    void lcs2() {
-        try {
-            String mset1 = jedis.mset("key1", "thisisasuperlongrandomtext", "key2", "isthisarandomlongtext");
-            assertEquals("OK", mset1);
-
-            LCSMatchResult lcs1 = jedis.strAlgoLCSKeys("key1", "key2", StrAlgoLCSParams.StrAlgoLCSParams().idx().withMatchLen());
-
-            assertEquals(lcs1.getLen(), 15);
-
-            assertEquals(lcs1.getMatches().get(0).getA().getStart(), 22);
-            assertEquals(lcs1.getMatches().get(0).getA().getEnd(), 25);
-            assertEquals(lcs1.getMatches().get(0).getB().getStart(), 17);
-            assertEquals(lcs1.getMatches().get(0).getB().getEnd(), 20);
-            assertEquals(lcs1.getMatches().get(0).getMatchLen(), 4);
-
-            assertEquals(lcs1.getMatches().get(1).getA().getStart(), 16);
-            assertEquals(lcs1.getMatches().get(1).getA().getEnd(), 21);
-            assertEquals(lcs1.getMatches().get(1).getB().getStart(), 7);
-            assertEquals(lcs1.getMatches().get(1).getB().getEnd(), 12);
-            assertEquals(lcs1.getMatches().get(1).getMatchLen(), 6);
-
-            assertEquals(lcs1.getMatches().get(2).getA().getStart(), 4);
-            assertEquals(lcs1.getMatches().get(2).getA().getEnd(), 6);
-            assertEquals(lcs1.getMatches().get(2).getB().getStart(), 4);
-            assertEquals(lcs1.getMatches().get(2).getB().getEnd(), 6);
-            assertEquals(lcs1.getMatches().get(2).getMatchLen(), 3);
-
-            assertEquals(lcs1.getMatches().get(3).getA().getStart(), 2);
-            assertEquals(lcs1.getMatches().get(3).getA().getEnd(), 3);
-            assertEquals(lcs1.getMatches().get(3).getB().getStart(), 0);
-            assertEquals(lcs1.getMatches().get(3).getB().getEnd(), 1);
-            assertEquals(lcs1.getMatches().get(3).getMatchLen(), 2);
-        } catch (Exception e) {
-            fail(e);
-        }
-    }
-
-    @Test
-    void substr() {
-        try {
-            String set = jedis.set("mykey", "This is a string");
-            assertEquals("OK", set);
-
-            String substr1 = jedis.substr("mykey", 0, 3);
-            assertEquals("This", substr1);
-
-            String substr2 = jedis.substr("mykey", -3, -1);
-            assertEquals("ing", substr2);
-
-            String substr3 = jedis.substr("mykey", -3, -2);
-            assertEquals("in", substr3);
-
-            String substr4 = jedis.substr("mykey", -3, -3);
-            assertEquals("i", substr4);
-
-            String substr5 = jedis.substr("mykey", 0, -1);
-            assertEquals("This is a string", substr5);
-
-            String substr6 = jedis.substr("mykey", 0, -2);
-            assertEquals("This is a strin", substr6);
-
-            String substr7 = jedis.substr("mykey", 10, 100);
-            assertEquals("string", substr7);
-        } catch (Exception e) {
-            fail(e);
-        }
-    }
-
-    @Test
-    void getex() {
-        try {
-            jedis.set("mykey", "Hello");
-//            String getex = jedis.getEx("mykey", GetExParams.getExParams()); // How to call getex without param??
-//            assertEquals("Hello", getex);
-//            long ttl = jedis.ttl("mykey");
-//            assertEquals(-1, ttl); // ttl unimplemented
-//            String getex = jedis.getEx("mykey", GetExParams.getExParams().ex(60));
-//            long ttl = jedis.ttl("mykey");
-//            assert(ttl <= 60);
-        } catch (Exception e) {
-            fail(e);
-        }
-    }
-
-    @Test
-    void msetnx() {
-        try {
-            long val1 = jedis.msetnx("key1", "Hello", "key2", "there");
-            assertEquals(1, val1);
-            long val2 = jedis.msetnx("key2", "new", "key3", "world");
-            assertEquals(0, val2);
-            List<String> values = jedis.mget("key1", "key2", "key3");
-            assertEquals("Hello", values.get(0));
-            assertEquals("there", values.get(1));
-            assertEquals(null, values.get(2));
-        } catch (Exception e) {
-            fail(e);
-        }
-    }
-
-    @Test
-    void psetnx() {
-        try {
-            String val1 = jedis.psetex("mykey", 1000, "Hello");
-            assertEquals("OK", val1);
-//            long val2 = jedis.pttl("mykey"); // pttl unimplemented
-//            assert(val2 <= 1000);
-        } catch (Exception e) {
-            fail(e);
-        }
-    }
-
-    @Test
-    void setex() {
-        try {
-            String val1 = jedis.setex("mykey", 10, "Hello");
-            assertEquals("OK", val1);
-//            long val2 = jedis.pttl("mykey"); // pttl unimplemented
-//            assert(val2 <= 1000);
-        } catch (Exception e) {
-            fail(e);
-        }
-    }
-
-    @Test
-    void setnx() {
-        try {
-            long val1 = jedis.setnx("mykey", "Hello");
-            assertEquals(1, val1);
-            long val2 = jedis.setnx("mykey", "World");
-            assertEquals(0, val2);
-            String val3 = jedis.get("mykey");
-            assertEquals("Hello", val3);
-        } catch (Exception e) {
-            fail(e);
-        }
-    }
-}
